third_party/WebKit/LayoutTests/external/wpt/async_cookies/document_cookie_static.tentative-expected.txt

This is a testharness.js-based test.
FAIL Cookie Store Tests (testDocumentCookie) assert_equals: testDeleteCookies failures: Error: assert_throws: Feature unavailable from unsecured contexts function "function () { throw e }" threw object "TypeError: Failed to execute 'delete' on 'CookieStore': __Secure- and __Host- cookies must be secure" ("TypeError") expected object "SyntaxError" ("SyntaxError"),Error: assert_throws: Feature unavailable from unsecured contexts function "function () { throw e }" threw object "TypeError: Failed to execute 'delete' on 'CookieStore': __Secure- and __Host- cookies must be secure" ("TypeError") expected object "SyntaxError" ("SyntaxError"),Error: assert_throws: Feature unavailable from unsecured contexts function "function () { throw e }" threw object "TypeError: Failed to execute 'delete' on 'CookieStore': __Secure- and __Host- cookies must be secure" ("TypeError") expected object "SyntaxError" ("SyntaxError"),Error: assert_throws: Feature unavailable from unsecured contexts function "function () { throw e }" threw object "TypeError: Failed to execute 'delete' on 'CookieStore': __Secure- and __Host- cookies must be secure" ("TypeError") expected object "SyntaxError" ("SyntaxError"),Error: assert_throws: Feature unavailable from unsecured contexts function "function () { throw e }" threw object "TypeError: Failed to execute 'delete' on 'CookieStore': __Secure- and __Host- cookies must be secure" ("TypeError") expected object "SyntaxError" ("SyntaxError"),Error: assert_throws: Feature unavailable from unsecured contexts function "function () { throw e }" threw object "TypeError: Failed to execute 'delete' on 'CookieStore': __Secure- and __Host- cookies must be secure" ("TypeError") expected object "SyntaxError" ("SyntaxError"),Error: assert_throws: Feature unavailable from unsecured contexts function "function () { throw e }" threw object "TypeError: Failed to execute 'delete' on 'CookieStore': __Secure- and __Host- cookies must be secure" ("TypeError") expected object "SyntaxError" ("SyntaxError") expected 0 but got 7
Harness: the test ran to completion.
